增加NopSpringTransactionFactory继承spring事务管理

+++ nop-dao/src/main/java/io/nop/dao/txn/ITransactionListener.java
@@ -19,7 +19,7 @@ public interface ITransactionListener extends IOrdered, Comparable<ITransactionL
     }
 
     default int compareTo(ITransactionListener o) {
-        return Double.compare(order(), o.order());
+        return Integer.compare(order(), o.order());
     }
 
     /**
